+++ example/lib/src/module/popup_page.dart
@@ -1,6 +1,7 @@
 import 'package:example/main.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import 'package:example/src/scaffold.dart';
 import 'package:fl_extended/fl_extended.dart';
 
 class PopupPage extends StatelessWidget {