+++ example/lib/src/module/refresh_page.dart
@@ -1,5 +1,6 @@
 import 'package:example/main.dart';
 import 'package:flutter/material.dart';
+import 'package:example/src/scaffold.dart';
 import 'package:fl_extended/fl_extended.dart';
 
 class _Item extends StatelessWidget {