+++ example/lib/src/module/scroll_list_page.dart
@@ -1,6 +1,7 @@
 import 'package:example/main.dart';
-import 'package:example/module/scroll_page.dart';
+import 'package:example/src/module/scroll_page.dart';
 import 'package:flutter/material.dart';
+import 'package:example/src/scaffold.dart';
 import 'package:fl_extended/fl_extended.dart';
 
 class ScrollListPage extends StatefulWidget {