+++ example/lib/src/module/text_field_page.dart
@@ -2,6 +2,7 @@ import 'package:example/main.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
+import 'package:example/src/scaffold.dart';
 import 'package:fl_extended/fl_extended.dart';
 
 class TextFieldPage extends StatelessWidget {