+++ example/lib/src/theme.dart
@@ -1,0 +1,11 @@
+import 'package:flutter/material.dart';
+
+class AppThemeData {
+  AppThemeData._();
+
+  static ThemeData get light =>
+      ThemeData(brightness: Brightness.light, useMaterial3: true);
+
+  static ThemeData get dark =>
+      ThemeData(brightness: Brightness.dark, useMaterial3: true);
+}