+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.2.0
+
+* Added the `UnifiedButton()`,It is based on several subclasses of `ButtonStyleButton`
+* Add the `unifiedButtonCategory` parameter to `Universal()`
+* Add some examples
+* Add a `marge` extension to `ThemeData`
+* Change the name of the `IconBox` `title` parameter to `label`
+
 ## 0.1.0
 
 * Split [flutter_waya](https://pub.dev/packages/flutter_waya)'s extension core into the current